requirements.txt
fastapi
uvicorn
twilio
openai
requests
pydub
sqlite3
python-dotenv
python-multipart